requirements.txt
numerize==0.12
aioredis<2.0
playwright==1.18.2
pymongo[srv]==4.0.1
cachetools==5.0.0
python-dotenv==0.19.2
git+https://github.com/paaksing/Pyot.git
git+https://github.com/DisnakeDev/disnake